% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_nhoods.R
\name{annotate_nhoods}
\alias{annotate_nhoods}
\title{Annotate Neighbourhoods with Majority Voting}
\usage{
annotate_nhoods(milo, cols_annot = c("celltype"))
}
\arguments{
\item{milo}{A Milo object with a filled nhoods slot.}

\item{cols_annot}{Character vector of column names from \code{colData} to use for annotation.}
}
\value{
A data.table with the following columns:
  \item{Nhood}{Neighbourhood identifier (numeric).}
  \item{Nhood_center}{Centre cell name of the neighbourhood.}
  For each annotation column, two columns are added:
  \item{<col>}{The winning annotation label for that neighbourhood.}
  \item{<col>.fraction}{The fraction of cells with the winning label.}
}
\description{
Performs majority voting annotation of neighbourhoods, where the most common label
among a neighbourhood's cells wins. This is a wrapper around \code{miloR::annotateNhoods}.
}
\details{
For each neighbourhood and each annotation column, the function identifies the most
frequent label among cells in that neighbourhood and records the fraction of cells
with the winning label.
}
\examples{
# Not run: annotate_nhoods(milo, cols_annot = c("celltype", "stage"))
}
