% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/internal-calculate_nhoodnhood_significance_oneway.R
\name{.calculate_nhoodnhood_significance_oneway}
\alias{.calculate_nhoodnhood_significance_oneway}
\title{Pipeline to give pvalue to nhood-nhood edges, oneway, not exposed to user}
\usage{
.calculate_nhoodnhood_significance_oneway(
  milos,
  name_intact = "milo_intact",
  name_scramble = "milo_scramble",
  dt_sims_true,
  col_scramble_label = "false",
  n_scrambles,
  assay = "logcounts",
  sim_method = "spearman",
  adjust,
  alpha_adjust
)
}
\arguments{
\item{milos}{A list of 2 Milo objects, the 1st one is kept intact, the 2nd one is scrambled.}

\item{name_intact, name_scramble}{Name given to the column with neighbourhoods from 1st (resp. 2nd) Milo object.}

\item{dt_sims_true}{A data.table with 3 columns. The first 2 columns indicate the pair of neighbourhoods, the third indicates the similarity value. This contains the similarity values between intact neighbourhoods.}

\item{col_scramble_label}{The name of a column in the 2nd Milo's colData. Label frequencies are estimated from this column and used in the scrambling process to weight the sampling probability of each cell.. Cells with most common samples are less likely to be sampled.}

\item{n_scrambles}{Number of rounds of cell resampling.}

\item{assay}{Assay from milos from which to calculate the neighbourhood expression.}

\item{sim_method}{Similarity metric. Must be one of c('pearson', 'kendall', 'spearman').}

\item{adjust}{Method to correct pvalues for multiple testing. If NULL does not do pvalue correction. Must be a valid method in 'p.adjust'.}

\item{alpha_adjust}{Level of significance to call significant edges. Float between 0 and 1.}
}
\description{
This function does the scrambling of nhoods, and calculates the pvalues against scrambled similarities. Runs only in one direction, i.e. one of the Milos is kept intact, the other is scrambled.
}
\details{
This function does:
1. Scramble neighbourhoods in the 2nd Milo, optionally taking labels to determine the resampling weights.
2. Calculate the similarities between true nhoods from the 1st Milo and the scrambled nhoods from the 2nd Milo
3. Calculate and adjust pvalues for each nhood-nhood edges by comparing dt_sims_true to the scrambled ones.
}
