% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-calculate_trimmed_ratios.R
\name{.calculate_trimmed_ratios}
\alias{.calculate_trimmed_ratios}
\title{Calculate the number of nhood-nhood edges that were trimmed after comparing true similarities to scrambled ones.}
\usage{
.calculate_trimmed_ratios(
  milos,
  dt_sims_withSignif,
  cols_label = c("celltype", "celltype")
)
}
\arguments{
\item{milos}{A list of 2 Milo objects.}

\item{dt_sims_withSignif}{A data.table with at least 3 columns. Each row represents a pair of nhoods across the Milos (column 1 and 2). A logical column named 'is_significant'}

\item{cols_label}{Vector of length 2, names of columns in the Milo's colData. Neighbourhoods will be annotated accroding to these columns and the number of (trimmed) edges will be calculated between the annotations.}
}
\value{
A data.table with 2 columns indicating the pair of neighbourhoods' labels, a column indication the total number of edges linking these pairs, a column with the number of significant edges, a column with the ratio of these 2.
}
\description{
Calculate the number of nhood-nhood edges that were trimmed after comparing true similarities to scrambled ones.
}
