% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-combine_2pvals_simes.R
\name{.combine_2pvals_simes}
\alias{.combine_2pvals_simes}
\title{Simes pvalue combination special case with 2 values}
\usage{
.combine_2pvals_simes(pval1, pval2)
}
\arguments{
\item{pval1, pval2}{The two pvalues to be combined.}
}
\value{
A single, combined pvalue.
}
\description{
Use non-generalizable Simes for speed up and ease of use in data.table
}
