% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-scramble_calculateWeights.R
\name{.scramble_calculateWeights}
\alias{.scramble_calculateWeights}
\title{Get the weights to randomly permute cell identity and create scrambled neighbourhoods}
\usage{
.scramble_calculateWeights(milo, col_scramble_label = "false")
}
\arguments{
\item{milo}{A Milo object with filled nhoods slot.}

\item{col_scramble_label}{The name of a column in the Milo's colData. Label frequencies are estimated from this column. If "false", returns NULL weights.}
}
\value{
A numerical vector of same length as the number of cells in milo. NULL if col_scramble_label=="false".
}
\description{
The weight of sampling each cell is based on the inverse frequency of its associated label.
}
