% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-weightedBipartiteGraphMatching.R
\name{.weightedBipartiteGraphMatching}
\alias{.weightedBipartiteGraphMatching}
\title{Weighted Bipartite Graph Matching}
\usage{
.weightedBipartiteGraphMatching(dt_signif_sims, col_sim = "sim")
}
\arguments{
\item{dt_signif_sims}{A data.table with 3 columns. The first 2 columns indicate the pair of neighbourhoods, the third indicates the similarity value. This should only contain the significant nhood-nhood edges to be used for matching.}

\item{col_sim}{Column name of dt_signif_sims containing the similarity values.}
}
\value{
A data.table where each row represent a pair of matched neighbourhoods.
}
\description{
Weighted Bipartite Graph Matching
}
