% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_centroid_nhoods.R
\name{extract_centroid_nhoods}
\alias{extract_centroid_nhoods}
\title{Extract neighbourhood coordinates from reduced dimension for visualisation}
\usage{
extract_centroid_nhoods(
  milo,
  dimred = "UMAP",
  use_indexCell = T,
  sel_dims = c(1, 2)
)
}
\arguments{
\item{milo}{A Milo object.}

\item{dimred}{The name of the reduced dimension.}

\item{use_indexCell}{Logical. If TRUE, use the coordinates of the neighbourhood index cells to represent the neighbourhoods. If FALSE, calculate the average coordinates of the individual cells (slower).}

\item{sel_dims}{Numeric of length 2. Which 2 coordinates to use.}
}
\value{
A data.table with 3 columns, indicating the nhood and its 2 coordinates in the reduced dimension.
}
\description{
Extract neighbourhood coordinates from reduced dimension for visualisation
}
