% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matches_map.R
\name{plot_matches_map}
\alias{plot_matches_map}
\title{Heatmap: what neighbourhoods were macthed together?}
\usage{
plot_matches_map(
  milos,
  dt_match,
  cols_label = c("celltype", "celltype"),
  cluster_rows = TRUE,
  cluster_cols = TRUE,
  transform = NULL
)
}
\arguments{
\item{milos}{A list of 2 Milo objects.}

\item{dt_match}{A data.table with at least 3 columns. Each row represents a pair of nhoods across the Milos (column 1 and 2). A logical column named 'is_significant'}

\item{cols_label}{Vector of length 2, names of columns in the Milo's colData. Neighbourhoods will be annotated accroding to these columns and the number of (trimmed) edges will be calculated between the annotations.}

\item{cluster_rows, cluster_cols}{Logical, whether to reorder the rows and columns with hierarchical clustering.}

\item{transform}{NULL or a character to transform the number of matches. Valid values include c("log10", "log1p", "sqrt"). See ?ggplot2::scale_fill_continuous for a full list.}
}
\value{
A ggplot object.
}
\description{
This function shows which neighbourhoods were matched by counting the number of matches linking neighbourhoods' annotations. The nhoods are first annotated, then the table of matches is checked against the annotations.
}
