% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trimming_map.R
\name{plot_trimming_map}
\alias{plot_trimming_map}
\title{Heatmap: how many edges linking were trimmed by RIMA?}
\usage{
plot_trimming_map(
  milos,
  dt_sims_withSignif,
  cols_label = c("celltype", "celltype"),
  cluster_rows = TRUE,
  cluster_cols = TRUE
)
}
\arguments{
\item{milos}{A list of 2 Milo objects.}

\item{dt_sims_withSignif}{A data.table with at least 3 columns. Each row represents a pair of nhoods across the Milos (column 1 and 2). A logical column named 'is_significant'}

\item{cols_label}{Vector of length 2, names of columns in the Milo's colData. Neighbourhoods will be annotated accroding to these columns and the number of (trimmed) edges will be calculated between the annotations.}

\item{cluster_rows, cluster_cols}{Logical, whether to reorder the rows and columns with hierarchical clustering.}
}
\value{
A ggplot object
}
\description{
This function shows how many nhood-nhood edges were trimmed by RIMA. The nhoods are first annotated to calculate the number of (un)significant edges linking 2 neighbourhoods' annotation.
}
