% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_nhoods.R
\name{annotate_nhoods}
\alias{annotate_nhoods}
\title{Annotate neighbourhoods}
\usage{
annotate_nhoods(milo, cols_annot = c("celltype"))
}
\arguments{
\item{milo}{A Milo object with a filled nhoods slot.}

\item{cols_annot}{The columns from colData to use for annotation.}
}
\value{
A data.table with the neighbourhoods' annotations. The two first
  columns contain a nhood label and the name of the nhood. Then a pair of
  column for each annotation indicating the winning label, and the fraction
  of cells in the neighbourhood with the winning label.
}
\description{
Do a majority voting annotations of neighbourhoods, i.e. the most label among
the neighbourhood's cells wins.
}
\details{
This is a wrapper around miloR::annotateNhoods
}
