% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_nhoodnhood_significance.R
\name{calculate_nhoodnhood_significance}
\alias{calculate_nhoodnhood_significance}
\title{Pipeline to give pvalue to nhood-nhood edges}
\usage{
calculate_nhoodnhood_significance(
  milos,
  dt_sims_true,
  col_scramble_label = "false",
  n_scrambles,
  sel_assay = "logcounts",
  sim_method = "spearman",
  adjust,
  alpha_adjust,
  direction = c("b", "lr", "rl")
)
}
\arguments{
\item{milos}{A list of 2 Milo objects, the 1st one is kept intact, the 2nd one is scrambled.}

\item{dt_sims_true}{A data.table with 3 columns. The first 2 columns indicate the pair of neighbourhoods, the third indicates the similarity value. This contains the similarity values between intact neighbourhoods.}

\item{col_scramble_label}{The name of a column in the 2nd Milo's colData. Label frequencies are estimated from this column and used in the scrambling process to weight the sampling probability of each cell.. Cells with most common samples are less likely to be sampled.}

\item{n_scrambles}{Number of rounds of cell resampling.}

\item{sel_assay}{Assay from milos from which to calculate the neighbourhood expression.}

\item{sim_method}{Similarity metric. Must be one of c('pearson', 'kendall', 'spearman').}

\item{adjust}{Method to correct pvalues for multiple testing. If NULL does not do pvalue correction. Must be a valid method in 'p.adjust'.}

\item{alpha_adjust}{Level of significance to call significant edges. Float between 0 and 1.}

\item{direction}{"Direction of the matching. Must be one of c('lr', 'rl', 'b')
If 'lr' ('lr'; short for left-right), nhoods in 1st Milo of milos (i.e. the left object) are kept intact and nhoods in 2nd Milo of milos (i.e. the right object) are scrambled to derive the significance of nhood-nhood similarities.
If 'rl' (right-left), this role of the 1st and 2nd Milo are inverted.
If 'b' (bidirectional), the nhood-nhood significance is calculated in both directions and combined with Simes' method.
The 'b' option is the most conservative and slower (because it runs both lr and rl) but accounts for the difference of heterogeneity between the 2 milos and is most reliable.",}
}
\value{
A data.table based on 'dt_sim_true', with 2 extra columns representing the pvalue and the adjusted pvalue associated with each similarity value.
}
\description{
This function takes in both Milos objects and the similarities between  does the scrambling of nhoods, and calculates the pvalues against scrambled similarities.
}
