% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_similarities.R
\name{calculate_similarities}
\alias{calculate_similarities}
\title{Calculate neighbourhood-neighbourhood similarities}
\usage{
calculate_similarities(milos, method = "spearman")
}
\arguments{
\item{milos}{List of 2 Milo objects with a filled nhoodExpression slot. Alternatively, can pass a list of 2 matrices representing neighbourhood expression.}

\item{method}{Similarity metric. Must be one of c('pearson', 'kendall', 'spearman').}
}
\value{
A data.table with 3 columns. The first 2 columns indicate the pair of neighbourhoods, the thrid indicate the similarity metric.
}
\description{
Calculate neighbourhood-neighbourhood similarities
}
