% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paired_expression.R
\name{get_paired_expression}
\alias{get_paired_expression}
\title{Plot paired expression}
\usage{
get_paired_expression(milos, dt_match, genes = NULL)
}
\arguments{
\item{milos}{List of 2 Milo objects with a filled nhoodExpression slot.}

\item{dt_match}{A data.table with 2 columns containing the nhood-nhood matching. The first (resp. second) column contains the names of nhoods in the 1st (resp. 2nd) Milo object.}

\item{genes}{A vector of gene names. The genes must be present in both Milo's rownames. If NULL, uses all available genes.}
}
\value{
A data.table
}
\description{
Plot paired expression
}
