% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_nhoods.R
\name{match_nhoods}
\alias{match_nhoods}
\title{Matches neighbourhoods with Weighted Bipartite Graph Matching}
\usage{
match_nhoods(
  dt_significant_nhoodnhood,
  filter_insignificant = TRUE,
  col_significance = "is_significant",
  return_sim = TRUE
)
}
\arguments{
\item{dt_significant_nhoodnhood}{A data.table with at least 3 columns as returned by 'calculate_nhoodnhood_significance'. Must contain a column named "sim". The 2 first columns must be contain the nhoods to be matched.}

\item{filter_insignificant}{Logical, whether to filter the rows in 'dt_significant_nhoodnhood', where the column 'col_significance' is not TRUE.}

\item{col_significance}{If filter_insignificant is TRUE; name of a logical column in dt_significant_nhoodnhood used to filter the rows where its value is TRUE.}

\item{return_sim}{Logical. If TRUE, add a 3rd column to the output that contains the similarity value of the nhood-nhood pair.}
}
\value{
A data.table with 2 columns containing the nhood-nhood matching.
}
\description{
Matches neighbourhoods with Weighted Bipartite Graph Matching
}
