% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matches_embed.R
\name{plot_matches_embed}
\alias{plot_matches_embed}
\title{Plot matches in side-by-side embeddings}
\usage{
plot_matches_embed(
  milos,
  dt_match,
  cols_color = c("celltype", "celltype"),
  dt_palette = NULL,
  dimred = "UMAP",
  args_process_coordinates = list(list(angle = 0, shift = c(0, 0)), list(angle = 0, shift
    = c(10, 0))),
  add_non_matched_category = c(),
  sim_limits = NULL,
  title = "",
  linewd = 0.35
)
}
\arguments{
\item{milos}{List of 2 Milo objects.}

\item{dt_match}{A data.table with 3 columns containing the nhood-nhood matching and similarities. Typically generated with match_nhoods.}

\item{cols_color}{A vector of length 2 containing the names of the columns in colData used to color the neighbourhoods. If NULL, no coloring.}

\item{dt_palette}{NULL or a data.table with 2 columns, one column must be named after cols_color and contains the neighbourhoods labels, the second column (named 'color') contains a HEX color code.}

\item{dimred}{Which reduced dimensions to use?}

\item{args_process_coordinates}{A list of 2 lists with preprocessing instructions for the neighbourhoods' coordinates. The nested lists must contain an entry called 'angle' with a scalar, and an entry called 'shift' with a numeric vector of length 2. The shift vector regulates the amount of horizontal and vertical shift between the 2 atlases to avoid overlapping. The angle is useful to rotate the atlases and disentangle matches lines.}

\item{add_non_matched_category}{not implemented yet}

\item{sim_limits}{Numeric vector of length 2. Limits the colorscale of the similarity.}

\item{title}{Title of the plot.}

\item{linewd}{Numeric, width of the matching lines}
}
\value{
A ggplot object.
}
\description{
Plot matches in side-by-side embeddings
}
