% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_paired_expression.R
\name{plot_paired_expression}
\alias{plot_paired_expression}
\title{Plot paired expression}
\usage{
plot_paired_expression(milos, dt_match, genes)
}
\arguments{
\item{milos}{List of 2 Milo objects with a filled nhoodExpression slot.}

\item{dt_match}{A data.table with 2 columns containing the nhood-nhood matching. The first (resp. second) column contains the names of nhoods in the 1st (resp. 2nd) Milo object.}

\item{genes}{A vector of genes names. The genes must be present in both Milo's rownames.}
}
\value{
A ggplot object, where each gene is plotted as a scatterplot in its own facet.
}
\description{
Plot paired expression
}
