% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_milos.R
\name{preprocess_milos}
\alias{preprocess_milos}
\title{Preprocess a pair of Milo objects for RIMA}
\usage{
preprocess_milos(
  milo1,
  milo2,
  assay = "logcounts",
  calculate_expression = TRUE,
  dt_features = NULL
)
}
\arguments{
\item{milo1, milo2}{Milo objects or path to RDS files with Milo objects. Their nhoods slot must be populated.
Rownames (i.e. features/gene names) must be specified to ensure that the right features
are compared; feature mapping between the two objects can be provided, see 'dt_features'.}

\item{assay}{Assay to calculate neighbourhoods' expression.}

\item{dt_features}{DataFrame with 2 columns that maps features across both milos in case the features are not fully overlapping.
The 1st (resp. 2nd) column refers to rownames in milo1 (resp. milo2).
This is useful for example to use orthologs to compare atlases of different species.
This can also be used to perform the matching on a subset of features.
If not specified, rownames must fully match between both milos.}

\item{calculate_similarities}{Whether to calculate neighbourhood's expression. Needed for downstream steps.}
}
\value{
A list of 2 Milo objects, with the same row names.
}
\description{
Preprocess a pair of Milo objects for RIMA
}
